+++ src/main/java/racingcar/controller/dto/RacingStartDTO.java
@@ -1,4 +1,4 @@
-package racingcar.service.dto;
+package racingcar.controller.dto;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
